refactor: refactor statement executor
