+++ src/frontend/Cargo.toml
@@ -21,12 +21,14 @@ chrono.workspace = true
 client = { workspace = true }
 common-base = { workspace = true }
 common-catalog = { workspace = true }
+common-config = { workspace = true }
 common-datasource = { workspace = true }
 common-error = { workspace = true }
 common-function = { workspace = true }
 common-grpc = { workspace = true }
 common-grpc-expr = { workspace = true }
 common-meta = { workspace = true }
+common-procedure = { workspace = true }
 common-query = { workspace = true }
 common-recordbatch = { workspace = true }
 common-runtime = { workspace = true }
@@ -42,7 +44,9 @@ futures = "0.3"
 futures-util.workspace = true
 humantime-serde.workspace = true
 itertools.workspace = true
+log-store = { workspace = true }
 meta-client = { workspace = true }
+raft-engine = { workspace = true }
 # Although it is not used, please do not delete it.
 meter-core.workspace = true
 meter-macros.workspace = true
