+++ src/log-store/src/raft_engine.rs
@@ -23,6 +23,9 @@ use crate::raft_engine::protos::logstore::{EntryImpl, NamespaceImpl};
 mod backend;
 pub mod log_store;
 
+pub use backend::RaftEngineBackend;
+pub use raft_engine::Config;
+
 pub mod protos {
     include!(concat!(env!("OUT_DIR"), concat!("/", "protos/", "mod.rs")));
 }
