+++ src/meta-client/src/client.rs
@@ -23,7 +23,7 @@ mod store;
 use api::v1::meta::Role;
 use common_error::ext::BoxedError;
 use common_grpc::channel_manager::{ChannelConfig, ChannelManager};
-use common_meta::ddl::{DdlExecutor, ExecutorContext};
+use common_meta::ddl::{DdlTaskExecutor, ExecutorContext};
 use common_meta::error::{self as meta_error, Result as MetaResult};
 use common_meta::rpc::ddl::{SubmitDdlTaskRequest, SubmitDdlTaskResponse};
 use common_meta::rpc::lock::{LockRequest, LockResponse, UnlockRequest};
@@ -178,7 +178,7 @@ pub struct MetaClient {
 }
 
 #[async_trait::async_trait]
-impl DdlExecutor for MetaClient {
+impl DdlTaskExecutor for MetaClient {
     async fn submit_ddl_task(
         &self,
         _ctx: &ExecutorContext,
